added a new button
that kind of works but not really we need to fix it

It's at the top of the window even though I told it to be at the bottom.
And the new attractions it creates don't show up until you click into an
existing attraction and back out.  In addition, having more than one
attraction with the same name (such as "New Attraction") breaks the save
logic, and the code will crash (AND LEAVE THE WINDOW OPEN) if you try to
close it.